Address -Wshorten-64-to-32

+++ src/common.c
@@ -61,7 +61,7 @@ long long numberise(char *str, int base)
 {
     char *p = str;
     // there are more efficient ways to strip '_' but this one is pretty clear
-    int len = strlen(p);
+    size_t len = strlen(p);
     while ((p = strchr(p, '_')))
         memmove(p, p + 1, len - (p - str));
