+++ src/os/Darwin/findself.c
@@ -7,7 +7,7 @@
 char *os_find_self(const char *argv0)
 {
     (void)argv0;
-    size_t size = PROC_PIDPATHINFO_MAXSIZE;
+    uint32_t size = PROC_PIDPATHINFO_MAXSIZE;
     char *path = malloc(size);
     if (proc_pidpath(getpid(), path, size) > 0) {
         return path;
